moving ext conversion into userData + updating test
